Extension to interact with login process
Use Docker to run latest p4d in a container, configured with a normal user
and a super user in a group without password timeout. Another container runs
the Node.js "auth service" application. The hook.js script is used to
install the Perforce extension and configure it for the containers.