Use long-polling to get user data
Repeated calls to the curl library in Lua can be problematic (in that it
explodes the process), so use one query that eventually times out on the
server. Also need to skip SSL verification since the auth service is likely
to have self-signed certs, and curl rejects those by default.