Support the previous SAML integration solution
The extension will detect the use of the previous SAML integration via the
auth.sso.args setting, in which case it conforms to the behavior of the SAML
triggers written in Python (found on the workshop). This relies on the auth
service, to which the desktop agent is directed for login, and also used to
validate the SAML response sent back by the agent.

For testing, the SAML agent does not like the self-signed certs in the auth
service, so the Docker Compose file can be changed to switch to http scheme
as needed. The default is to use https for the new integration solution.