Use unique identifier for each new login request
Since a user might log in to multiple client systems, the login status
should be keyed by a unique request identifier. This, like the user status,
is cached in memory for a short time (on the order of minutes). Also, by
using a unique identifier, any erroneous queries for user status will return
a 404 immediately rather than waiting a minute to get a 408.

Moved the duplicate code from oidc and saml routes to a new module.