Define container for running the extension
Uses a named network that is also defined in the "application" on the auth
service side, allowing the containers to find each other.

Unlike previous iterations, this container includes a build process that
packages, installs, and configures the extension to connect to the auth
service running in the other container.