AuthMethod ldap users will not use SSO login
It makes no sense for a Perforce user with AuthMethod set to 'ldap' to be
logging in using the SSO mechanism (extension + service). As such, the
extension now automatically defers for users that authenticate with LDAP. By
deferring, the extension passes the authentication responsibility back to
the server, which will then use LDAP (or password if that fallback is
allowed via configure settings).

This saves the administrator from having to perform double bookkeeping,
setting the AuthMethod _and_ adding the user to the non-sso-users/groups
configuration in the extension.