Use the "old" P4 API consistently for now
Seems that the new API is what should be used, but before doing that, change
both of the functions using the API to use the previous form of the API.
Both will be switched to the preferred API usage at a later time.

Also, make the functions consistent in terms of returning an "ok" that
indicates if there was a problem in the API calls.