Compare user identitiers case-insensitively
An issue came up with a customer in which the authentication failed because
the email addresses in the Perforce user specs did not match in letter-case
with the email addresses in the identity provider. This change now forces
both the "user-identifier" and "name-identifier" values to be lowercased
before comparing.

If this becomes a problem, then the extension can be modified to take the
server case-sensitivity setting into account, or add an extension setting
just for this purpose.