Use SSL curl options only when using HTTPS
When connecting to the authenticate service, add the libcurl SSL options
only if the Service-URL is using the https protocol. Otherwise, the SSL
options are not applied, as that would not make sense. When the service is
configured to use plain HTTP, it disables the requirement for client certs
in retrieving the user profile data. This is not recommended, but may be
necessary depending on the network architecture.

npm test passes