feat: new configure script for login extension
Similar to the configure script for Helix Core, this Linux-only bash shell
script will assist in the installation and configuration of the loginhook
extension. The script can be run interactively or via command line
arguments, and will prompt to restart p4d after the extension is installed.
When run non-interactively, it will restart p4d only if the --yes option is
provided.