fix: remove -- and -e from awk invocation
The configure script had redundant arguments to awk that seemingly caused a
problem on some systems. These arguments are documented and supported for
the version of awk on these systems, but perhaps some other version of awk
is getting in the way. Maybe by removing these arguments the issue will be
resolved.