Remove platform check from extension config script
The configure script for the extension does not require a specific version
of Linux. In fact, any Unix-like system should suffice, including macOS, so
checking for Debian vs RedHat is only a hinderance to broader support.

Added a note to the internal docs about how to run the script on macOS.