Add more logging when request for identifier fails
When the HTTP request to HAS for the login identifier fails for any reason,
write an additional log entry suggesting that the administrator double-check
the Service-URL value (the log entry contains the Service-URL value as well).
If this URL is invalid for any reason, the errors reported by libcurl can be
less than helpful.

Would have displayed a message to the client, but sending client messages
from the auth-pre-sso hook appears to be unsupported. For now, logging is
the only option available to the extension.

npm test passes