feat: add sso-groups extension configuration field
Essentially the opposite of non-sso-groups, the sso-groups list names
Perforce groups whose members _must_ authenticate using the extension. That
is, when this field is defined, any user of any group named in this field
must login using SSO, while any user that is _not_ a member of any group
listed will be excluded from SSO authentication. When this field contains a
list of groups then the non-sso-groups field will be ignored.

npm test passes