build: combine apt-get update and install into one step
Sometimes the automated builds would fail because the package cache updated
with `apt-get update` was cached by docker, but the online repository had
been updated since the last successful build. As a result, the subsequent
step in the docker build, `apt-get install`, would fail with a "404 not
found" on the outdated packages. By combining the two commands into a single
step in the docker build, we should avoid this problem.