docker: patch p4d configure script to work reliably
Seems that the bash script for configuring p4d has a timing bug, such that
after starting p4d, there is a chance it will "fail" to start. The issue is
that the bash script runs a tight loop 10 times trying to connect to p4d,
but this is very prone to variations in process timing. The loop could get
all of the CPU cycles and spin out before p4d has a chance to start. Adding
a sleep of 1 second to the loop fixes the problem.