fix: capture messages from p4 cmds to the log
When the extension runs p4 commands, such as `p4 user` and `p4 groups`, make
sure to capture any informative messages or errors and write them to the
logs. Normally nothing is catpured, but if an error occurs, we may never
know about it without this additional reporting.

Also, when building the auto client API instance, check for any error that
might be returned, and if there is an error, report it to the log and fail
out of the function.

npm test passes