fix: make configure script inputs consistent
Some of the configure script inputs where using the bash select function,
which meant that they could not display and accept the default value without
the user making a selection. This affected both the "enable logging" and
"user identifier" inputs; now they are like all other inputs, a free-form
text input that displays and accepts a default value.