feat: add settings to change paths to client certs
The extension uses a client certificate to connect to HAS, and that has so
far been a pair of files named client.crt and client.key, found in the
extension directory. Likewise, the CA certificate to verify the service is
given in the ca.crt file. With this change, three new global extension
settings allow the admin to specify paths to a different set of files. They
are named Client-Cert, Client-Key, and Authority-Cert, and are documented in
the admin guide. The previous default files will be in effect as long as the
new settings start with "...", similar to the other extension settings.

npm test passes