fix: do not fall through 1-step validation
The extension validation logic has two paths, one for the so-called "1-step"
process used by Swarm, and the other for the 2-step "normal" login process
used by every other client (p4, p4v, etc). If the 1-step validation
encounters an error, it logs the error and falls through to the 2-step
validation process. This works because technically the user authenticated
successfully, but it hides the fact that there is an error occurring. The
appropriate behavior is to fail when there is an error during the intended
validation process, not bypass that and do something else.

npm test passes