fix: improve summary of extension configuration
Near the end of the HAS extension configuration, the output can get a little
strange if the authentication protocol value is left undefined. Similarly,
the logging flag also looks strange if left alone. This change fixes both so
that the summary display just before making the extension config changes is
easier to understand.