fix: prompt to allow password for non-SSO users
When configuring the extension and the user enters one or more users or
groups that will be not using SSO, then prompt the user to determine if they
would like the script to set the server configurable to allow passwords.
Otherwise you install and configure the extension but no one can
authenticate without using SSO, which is not recommended for admin users.