fix: use new option to install unsigned extension
Since version 2020.2/1943129 the Helix Core Server supports signed
extensions. As a result, it requires extensions to be signed in order to
install them, without either setting a configurable or passing a new client
command option. This change adds the use of that flag, if the server version
indicates signed extension support. This is en lieu of having a signed
extension at the time of installation, which is not possible when fetching
the extension code from GitHub (would need the private signing key).