fix: add default for Service-URL to avoid lock-out
If the loginhook extension configurtion does not have Service-URL at
all (that is, the user deleted that entry from the configuration), then the
extension fails catastrophically such that no user can authenticate. All
other settings in the extension have defaults that may or may not work, but
at least the extension still functions enough to allow non-SSO accounts to
authenticate, if the server is configured as such.

As such, have the extension provide a simplistic default for Service-URL,
which likely won't have any functional value, but at least the extension
will not fail all logins.

npm test passes