fix: add instanceId to URLs for login and status
The extension was failing to include the instanceId query parameter when
constructing the various URLs used to connect to the service. This bug would
defeat the rule-based routing support that HAS offers through this query
parameter. This change fixes that issue, now all login URLs, status URLs,
and the SAML validation URL will include the instanceId parameter. Note that
the OAuth validation URL does not need this parameter as JSON Web Token
based authentication is completely stateless.

npm test passes