fix: if service unreachable, show helpful message
Previously if the extension was unable to connect to the auth service, it
would silently fail out and p4d would invoke another authentication method.
This was confusing at best, as the admin needed to know to check the debug
logs, assuming they were enabled.

Now, if the extension has an error when connecting to the auth service, it
will fall through to the auth-check hook where it can send the message to
the client. Note that the pre-check hook cannot send message to the client,
which is why errors in pre-check are difficult to report. Falling through in
this manner also means the user cannot authenticate, which is no different
than before.

Note that other failure conditions are not treated in this way, as they
occur too early in the process to be certain that the user is expected to
authenticate using SSO. To avoid locking out administrators and other
service users, the extension will continue to fail out as before.

npm test passes