fix: fail fast when extension lacks a valid ticket
If the extension gets an "info" message indicating that the ExtP4USER
invoking commands against the server is missing a valid ticket, then log a
clear error message and fail out. This problem is otherwise difficult to
diagnose and can go undetected for too long.

The effect of this change is that the extension will not optimistically
proceed through the rest of the authentication process, hoping that by some
chance it will happen to work out. In fact, that has not really ever been
the case, as the lack of a valid p4 ticket prevents properly categorizing
the user into either the "sso required" or "sso excluded" groups.

npm test passes