feat: add command to test extension before restart
The extension now implements the RunCommand entry point, allowing for the
testing of the extension between the time of installing the extension and
restarting the server. This is helpful to ensure that basic functionality is
working before "flipping the switch" and potentially locking out users.

The new commands are test-svc, test-cmd, and test-all, which test the
connection to the authentication service, server command invocation, and
both, respectively.

npm test passes