fix: build new extension before removing old one
The configure script was first removing the existing extension and then
building the new extension. Normally this is not a problem, but if the p4
client is too old to support the extension --package command, then the
server will be left in a broken state until the admin can update the p4
client and complete the installation.

With this change, the configure script will attempt to build the extension
first, and if that fails, exit immediately before making any other changes.