fix: return url for P4API clients that can use it
It seems that some (or maybe just one for the time being) can use the data
returned from the loginhook extension. In an earlier change the url was
removed from the return values for P4PHP and the HTH clients, as it seemed
that the value (named "data" in the p4d login code) was not being used.
However, a change in the P4Ruby client seems to make use of that value, and
as such the HTH client no longer has access to the login URL.

This change reverses that, and provides comments that explain how the values
might be used by P4API clients.