feat: add configurable "service down" URL
If the authentication service is not reachable from the extension, the user
will be sent to "http://example.com" and an error message will be shown on
the client. However, going to example.com is not especially helpful, so the
extension now has a new global setting named Service-Down-URL that allows
the admin to change the value from example.com to another URL that might be
more helpful to the user (such as an IT request page).

npm test passes