support: fix info gathering script based on input
Fixed the problem of the disappearing tar file (it was being created in the
directory containing the extension installation files) and provide a
description of the commands being invoked as information is collected. The
script finishes in a second, so this is all after the fact, but at least it
gives some clue what the script was doing.