config: retain extra settings during upgrade
When running the extension configure script the settings that are normally
not managed by the configure script would be lost during an upgrade. That
is, if an extension is already installed and has the Service-Down-URL
settings, for example, that setting would be lost when the configure script
is run to install a newer version of the extension. This change corrects
that, and all known settings are preserved during an upgrade.