fix: new Resolve-Host for weird DNS situations
In certain circumstances, it may be that the p4d container needs to connect
to the auth service using an IP address rather than a host name. However,
doing so results in other problems, like libcurl deciding not to send the
client certificate. To rectify this, a new setting name Resolve-Host has
been added that allows for mapping the hostname to a specific port and IP
address. This acts as a simplistic DNS lookup and allows libcurl to send the
client certs as expected. By supplying the host name the load balancer
and/or reverse proxy can route the request to the appropriate backend
service.

npm test passes