fix: try ssl in configure during p4 prompting
Change the extension configure script to try "ssl:" on the P4PORT if the
initial input by the user does not work, and the P4PORT did not already have
an 'ssl' prefix. This is a fairly common mistake: p4d is configured for SSL
but the admin running the configure script forgets and inputs a port value
like "1666" or "localhost:1666" -- all that is needed is for the script to
try the "ssl:" prefix.