doc: example k8s setup using Google Cloud
The README and manifests under kubernetes/google offer an easier alternative
to getting Helix Core Server running in Kubernetes in the cloud. The original
guide has been moved to kubernetes/kubeadm to reflect its focus on setting
up a local cluster using the kubeadm tool.

Note that running Helix Core Server in Kubernetes is still a terrible idea,
as discussed at the top of the README in each directory.