test: remove patch for p4d configure delay bug
The p4d configure script had a bug in which it looped to wait for "slow"
systems to start p4d, except that there was no delay in the loop and hence
it was not effective at all. This bug has subsequently been fixed in at
least 2023.2 and this change removes the hack for dealing with it.