fix: print warning if super ticket expires soon
The extension will behave incorrectly if the ticket for the ExtP4USER
expires, and the value for ExtP4USER is the super user that is running the
configure script. As such, the configure script should check that the ticket
expires a long time into the future.

With this change, if the super user running the configure script has a
ticket that expires in less than 30 days, a warning will be printed in
yellow that calls attention to the problem.