fix: avoid stale tickets issue
Use the work-around described in the extension documentation to avoid the
problem of p4d serving the extension a stale ticket. That is, if the tickets
on disk are stale, p4d will give that to the extension. This results in the
extension being unable to invoke commands, which affects the group-based
filtering. This fix removes that problem such that it ignores the tickets
file entirely and p4d will always use a new generated ticket every time.

npm test passes