doc: add example code for disabling extension
In the event that everyone is locked out of P4 Server due to a misconfigured
extension, an administrator can edit the extension source code and uncomment
the line in the AuthPreSSO function to temporarily disable the extension and
let users access P4 Server via non-SSO authentication. This only helps if
there are admin accounts that are not using SSO for authentication, however.

Look for the comment after the "function AuthPreSSO()" line in the main.lua
file under 117E9283-732B-45A6-9993-AE64C354F1C5/1-arch of the extensions
directory under the server root.