fix: log p4 errors to the extension log
After the fix for working around the stale ticket issue, logging of p4
command failures was (unintentionally) removed. As such, any time a p4
command fails there will be nothing in the extension logs. This fixes that
mistake such that any p4 errors will appear in the logs, if logging is
enabled with the enable-logging extension setting.

npm test passes