feat: allow using ticket for configure script
When running the extension configure script, it is now possible to enter the
super user's ticket value rather than a password. Some times an admin will
enter the password hoping that it would work; with this change that will
work as expected. Entering a password still works.

The pattern match is anything that is exactly 32 alphanumeric characters
which is typically a ticket value. Even if the value entered was not a
ticket, the login status will be verified and the script should proceed
normally. If, however, the admin does not already have a valid ticket and
the password they entered looks like a ticket, then this will fail and the
admin will have to login separately and provide the actual ticket value when
invoking the configure script. Hopefully that scenario will affect a very
limited number of users, if any.